Upgrade Stack resolver from LTS-7.x to LTS-12.x
Unfortunately, the "long-term" part of "LTS" has turned out to be a
lie.  Only 3 years in, and there is now a regression that causes Stack
to fail to build Alex and Happy (commercialhaskell/stack#5242).  There
is no easy fix, so I have upgraded the Stack resolver to 12.26, which
seems to work better.  I hope that this one is supported longer.

As part of this fix, I weakened the version requirements in the Cabal
package specification.  Everything still works as far as I can tell,
but it won't hold forever: LTS-14.x and higher have problems because of
the MonadFail changes.  Therefore the version specifiers are not a
true reflection of the package requirements.